refactor(textlint): use New CLI by default
- remove old CLI from source code
- export New CLI as `cli`
