+++ packages/textlint/package.json
@@ -25,8 +25,7 @@
     "main": "lib/src/index.js",
     "types": "lib/src/index.d.ts",
     "bin": {
-        "textlint": "./bin/textlint.js",
-        "textlint-esm": "./bin/textlint-esm.js"
+        "textlint": "./bin/textlint.js"
     },
     "directories": {
         "test": "test/"
