--- packages/textlint/test/cli-new/fixtures/.textlintrc.json
@@ -1,7 +1,0 @@
-{
-  "rules": {
-    "textlint-rule-no-todo": {
-      "severity": "warning"
-    }
-  }
-}
