--- packages/textlint/test/cli-new/fixtures/fixer-rules/input.md
@@ -1,1 +1,0 @@
-This is fix<REMOVE_MARK>
