--- packages/textlint/test/cli-new/fixtures/fixer-rules/output.md
@@ -1,1 +1,0 @@
-This is fixed.
