--- packages/textlint/test/cli-new/fixtures/pass.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-OK
-</body>
-</html>
