--- packages/textlint/test/cli-new/fixtures/quite.textlintrc.json
@@ -1,10 +1,0 @@
-{
-  "plugins": [
-    "html"
-  ],
-  "rules": {
-    "no-todo": {
-      "severity": "warning"
-    }
-  }
-}
