--- packages/textlint/test/cli-new/fixtures/rules/no-error/textlint-rule-no-error.ts
@@ -1,6 +1,0 @@
-import { TextlintRuleContext } from "@textlint/types";
-
-// Do no error rule
-export default function (_context: TextlintRuleContext) {
-    return {};
-}
