--- packages/textlint/test/cli-new/fixtures/test.md
@@ -1,1 +1,0 @@
-test.
