--- packages/textlint/test/cli-new/fixtures/todo.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<p>TODO: THIS IS TODO</p>
-</body>
-</html>
