--- packages/textlint/test/cli-new/fixtures/todo.md
@@ -1,3 +1,0 @@
-This is text.
-
-- [ ] TODO
