+++ packages/textlint/test/cli/cli-test.ts
@@ -1,9 +1,11 @@
 // LICENSE : MIT
 "use strict";
 import * as assert from "assert";
-import { cli } from "../../src";
+import { cli } from "../../src/cli";
 import * as path from "path";
 import { Logger } from "../../src/util/logger";
+import fs from "fs";
+import { loadBuiltinPlugins } from "../../src/loader/TextlintrcLoader";
 
 type RunContext = {
     getLogs(): string[];
@@ -32,7 +34,13 @@ const runWithMockLog = async (cb: (context: RunContext) => unknown): Promise<unk
             assert.ok(messages.length === 0, "should not have logs");
         }
     };
-    await cb(context);
+    try {
+        await cb(context);
+    } catch (error) {
+        // eslint-disable-next-line no-console
+        console.log("Logs", context.getLogs());
+        throw error;
+    }
     Logger.log = originLog;
     return;
 };
@@ -63,7 +71,7 @@ describe("cli-test", function () {
         it("should return error when text with incorrect quotes is passed as argument", function () {
             return runWithMockLog(async () => {
                 const ruleDir = path.join(__dirname, "fixtures/rules");
-                const result = await cli.execute(`--rulesdir ${ruleDir}`, "text");
+                const result = await cli.execute(`--rulesdir ${ruleDir} --stdin-filename=test.md`, "text");
                 assert.strictEqual(result, 1);
             });
         });
@@ -75,14 +83,14 @@ describe("cli-test", function () {
                 assert.strictEqual(result, 1);
             });
         });
-        it("should return an error because No rules found", function () {
+        it("should return an error because No rules found with text", function () {
             return runWithMockLog(async ({ getLogs }) => {
-                const result = await cli.execute("", "text");
+                const result = await cli.execute("--stdin-filename=test.md", "text");
                 assert.strictEqual(result, 1);
                 assert.match(getLogs()[0], /No rules found/);
             });
         });
-        it("should return an error because No rules found", function () {
+        it("should return an error because No rules found with files", function () {
             return runWithMockLog(async ({ getLogs }) => {
                 const targetFile = path.join(__dirname, "fixtures/test.md");
                 const result = await cli.execute(`${targetFile}`);
@@ -118,7 +126,7 @@ describe("cli-test", function () {
                                 },
                                 range: [17, 18],
                                 message: "Found TODO: '- [ ] TODO'",
-                                ruleId: "no-todo",
+                                ruleId: "textlint-rule-no-todo",
                                 severity: 1,
                                 type: "lint"
                             }
@@ -290,6 +298,98 @@ describe("cli-test", function () {
             const result = await cli.execute(`${targetFile}`);
             assert.strictEqual(result, 1);
             assertMatchLog(/No rules found/);
+        });
+    });
+    describe("--cache", function () {
+        const cacheLocation = path.join(__dirname, ".textlintcache");
+        afterEach(() => {
+            try {
+                fs.unlinkSync(cacheLocation);
+            } catch {
+                // nope
+            }
+        });
+        it("should cache the result and create .textlintcache", function () {
+            return runWithMockLog(async () => {
+                const targetFile = path.join(__dirname, "fixtures/test.md");
+                const ruleModuleName = "textlint-rule-no-todo";
+                const result = await cli.execute(
+                    `--cache --cache-location "${cacheLocation}" --rule "${ruleModuleName}" ${targetFile}`
+                );
+                assert.strictEqual(fs.existsSync(cacheLocation), true);
+                assert.strictEqual(result, 0);
+            });
+        });
+    });
+    describe("--print-config", function () {
+        it("should print current descriptor.toJSON()", function () {
+            return runWithMockLog(async ({ getLogs }) => {
+                const result = await cli.execute(`--print-config`);
+                assert.strictEqual(result, 0);
+                const expected = await loadBuiltinPlugins();
+                assert.strictEqual(getLogs()[0], JSON.stringify(expected.toJSON(), null, 4));
+            });
+        });
+        it("should print current descriptor.toJSON() with .textlintrc", function () {
+            return runWithMockLog(async ({ getLogs }) => {
+                const configFile = path.join(__dirname, "fixtures/.textlintrc.json");
+                const result = await cli.execute(`--print-config --config ${configFile}`);
+                assert.strictEqual(result, 0);
+                const expected = await loadBuiltinPlugins();
+                assert.strictEqual(
+                    getLogs()[0],
+                    JSON.stringify(
+                        {
+                            ...expected.toJSON(),
+                            rule: [
+                                {
+                                    ruleId: "textlint-rule-no-todo",
+                                    options: {
+                                        severity: "warning"
+                                    }
+                                }
+                            ],
+                            configBaseDir: path.dirname(configFile)
+                        },
+                        null,
+                        4
+                    )
+                );
+            });
+        });
+        it("should print current descriptor.toJSON() when add a --rule", function () {
+            return runWithMockLog(async ({ getLogs }) => {
+                const ruleModuleName = "textlint-rule-no-todo";
+                const result = await cli.execute(`--print-config --rule ${ruleModuleName}`);
+                assert.strictEqual(result, 0);
+                const expected = await loadBuiltinPlugins();
+                assert.strictEqual(
+                    getLogs()[0],
+                    JSON.stringify(
+                        {
+                            ...expected.toJSON(),
+                            rule: [
+                                {
+                                    ruleId: "no-todo",
+                                    options: true
+                                }
+                            ]
+                        },
+                        null,
+                        4
+                    )
+                );
+            });
+        });
+    });
+    describe("--no-textlint", function () {
+        it("should not load textlintrc, but load built-in plugins", function () {
+            return runWithMockLog(async ({ getLogs }) => {
+                const result = await cli.execute(`--no-textlintrc --print-config`);
+                assert.strictEqual(result, 0);
+                const expected = await loadBuiltinPlugins();
+                assert.strictEqual(getLogs()[0], JSON.stringify(expected.toJSON(), null, 4));
+            });
         });
     });
     describe("--version", function () {
